Change TextFields to Material3
