+++ core/ui/compose/designsystem/src/main/kotlin/app/k9mail/core/ui/compose/designsystem/atom/textfield/TextFieldLabel.kt
@@ -1,0 +1,20 @@
+package app.k9mail.core.ui.compose.designsystem.atom.textfield
+
+import androidx.compose.runtime.Composable
+import androidx.compose.material3.Text as Material3Text
+
+private const val ASTERISK = "*"
+
+@Composable
+internal fun TextFieldLabel(
+    label: String,
+    isRequired: Boolean,
+) {
+    Material3Text(
+        text = if (isRequired) {
+            "$label$ASTERISK"
+        } else {
+            label
+        },
+    )
+}
