+++ core/ui/compose/designsystem/src/main/kotlin/app/k9mail/core/ui/compose/designsystem/atom/textfield/TextFieldOutlined.kt
@@ -3,11 +3,7 @@ package app.k9mail.core.ui.compose.designsystem.atom.textfield
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import app.k9mail.core.ui.compose.designsystem.atom.icon.Icon
-import app.k9mail.core.ui.compose.designsystem.atom.icon.Icons
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
-import androidx.compose.material.OutlinedTextField as MaterialOutlinedTextField
+import androidx.compose.material3.OutlinedTextField as Material3OutlinedTextField
 
 @Suppress("LongParameterList")
 @Composable
@@ -24,7 +20,7 @@ fun TextFieldOutlined(
     isSingleLine: Boolean = true,
     keyboardOptions: KeyboardOptions = KeyboardOptions.Default,
 ) {
-    MaterialOutlinedTextField(
+    Material3OutlinedTextField(
         value = value,
         onValueChange = if (isSingleLine) stripLineBreaks(onValueChange) else onValueChange,
         modifier = modifier,
@@ -36,77 +32,4 @@ fun TextFieldOutlined(
         singleLine = isSingleLine,
         keyboardOptions = keyboardOptions,
     )
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun TextFieldOutlinedPreview() {
-    PreviewWithThemes {
-        TextFieldOutlined(
-            value = "Input text",
-            onValueChange = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun TextFieldOutlinedWithLabelPreview() {
-    PreviewWithThemes {
-        TextFieldOutlined(
-            value = "Input text",
-            onValueChange = {},
-            label = "Label",
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun TextFieldOutlinedDisabledPreview() {
-    PreviewWithThemes {
-        TextFieldOutlined(
-            value = "Input text",
-            onValueChange = {},
-            isEnabled = false,
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun TextFieldOutlinedErrorPreview() {
-    PreviewWithThemes {
-        TextFieldOutlined(
-            value = "Input text",
-            onValueChange = {},
-            hasError = true,
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun TextFieldOutlinedRequiredPreview() {
-    PreviewWithThemes {
-        TextFieldOutlined(
-            value = "",
-            onValueChange = {},
-            label = "Label",
-            isRequired = true,
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun TextFieldOutlinedWithTrailingIconPreview() {
-    PreviewWithThemes {
-        TextFieldOutlined(
-            value = "",
-            onValueChange = {},
-            trailingIcon = { Icon(imageVector = Icons.Filled.user) },
-            isRequired = true,
-        )
-    }
 }
