+++ core/ui/compose/designsystem/src/main/kotlin/app/k9mail/core/ui/compose/designsystem/atom/textfield/TextFieldOutlinedEmailAddress.kt
@@ -4,9 +4,7 @@ import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.tooling.preview.Preview
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
-import androidx.compose.material.OutlinedTextField as MaterialOutlinedTextField
+import androidx.compose.material3.OutlinedTextField as Material3OutlinedTextField
 
 @Suppress("LongParameterList")
 @Composable
@@ -20,7 +18,7 @@ fun TextFieldOutlinedEmailAddress(
     isRequired: Boolean = false,
     hasError: Boolean = false,
 ) {
-    MaterialOutlinedTextField(
+    Material3OutlinedTextField(
         value = value,
         onValueChange = stripLineBreaks(onValueChange),
         modifier = modifier,
@@ -34,51 +32,4 @@ fun TextFieldOutlinedEmailAddress(
         ),
         singleLine = true,
     )
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun TextFieldOutlinedEmailAddressPreview() {
-    PreviewWithThemes {
-        TextFieldOutlinedEmailAddress(
-            value = "Input text",
-            onValueChange = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun TextFieldOutlinedEmailAddressWithLabelPreview() {
-    PreviewWithThemes {
-        TextFieldOutlinedEmailAddress(
-            value = "Input text",
-            label = "Label",
-            onValueChange = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun TextFieldOutlinedEmailDisabledPreview() {
-    PreviewWithThemes {
-        TextFieldOutlinedEmailAddress(
-            value = "Input text",
-            onValueChange = {},
-            isEnabled = false,
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun TextFieldOutlinedEmailErrorPreview() {
-    PreviewWithThemes {
-        TextFieldOutlinedEmailAddress(
-            value = "Input text",
-            onValueChange = {},
-            hasError = true,
-        )
-    }
 }
