+++ core/ui/compose/designsystem/src/main/kotlin/app/k9mail/core/ui/compose/designsystem/atom/textfield/TextFieldOutlinedFakeSelect.kt
@@ -3,16 +3,14 @@ package app.k9mail.core.ui.compose.designsystem.atom.textfield
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.interaction.PressInteraction
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import app.k9mail.core.ui.compose.designsystem.atom.icon.Icon
 import app.k9mail.core.ui.compose.designsystem.atom.icon.Icons
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
-import androidx.compose.material.OutlinedTextField as MaterialOutlinedTextField
+import androidx.compose.material3.OutlinedTextField as Material3OutlinedTextField
+import androidx.compose.material3.Text as Material3Text
 
 @Composable
 fun TextFieldOutlinedFakeSelect(
@@ -21,7 +19,7 @@ fun TextFieldOutlinedFakeSelect(
     modifier: Modifier = Modifier,
     label: String? = null,
 ) {
-    MaterialOutlinedTextField(
+    Material3OutlinedTextField(
         value = text,
         onValueChange = { },
         modifier = Modifier
@@ -44,27 +42,4 @@ fun TextFieldOutlinedFakeSelect(
     )
 }
 
-private fun optionalLabel(label: String?): @Composable (() -> Unit)? = label?.let { { Text(label) } }
-
-@Preview
-@Composable
-internal fun TextFieldOutlinedFakeSelectPreview() {
-    PreviewWithThemes {
-        TextFieldOutlinedFakeSelect(
-            text = "Current value",
-            onClick = {},
-        )
-    }
-}
-
-@Preview
-@Composable
-internal fun TextFieldOutlinedFakeSelectPreviewWithLabel() {
-    PreviewWithThemes {
-        TextFieldOutlinedFakeSelect(
-            text = "Current value",
-            onClick = {},
-            label = "Label",
-        )
-    }
-}
+private fun optionalLabel(label: String?): @Composable (() -> Unit)? = label?.let { { Material3Text(label) } }
