+++ core/ui/compose/designsystem/src/main/kotlin/app/k9mail/core/ui/compose/designsystem/atom/textfield/TextFieldOutlinedPassword.kt
@@ -1,8 +1,6 @@
 package app.k9mail.core.ui.compose.designsystem.atom.textfield
 
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -13,11 +11,11 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.tooling.preview.Preview
 import app.k9mail.core.ui.compose.designsystem.R
 import app.k9mail.core.ui.compose.designsystem.atom.icon.Icons
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
-import androidx.compose.material.OutlinedTextField as MaterialOutlinedTextField
+import androidx.compose.material3.Icon as Material3Icon
+import androidx.compose.material3.IconButton as Material3IconButton
+import androidx.compose.material3.OutlinedTextField as Material3OutlinedTextField
 
 @Suppress("LongParameterList")
 @Composable
@@ -33,7 +31,7 @@ fun TextFieldOutlinedPassword(
 ) {
     var passwordVisibilityState by rememberSaveable { mutableStateOf(false) }
 
-    MaterialOutlinedTextField(
+    Material3OutlinedTextField(
         value = value,
         onValueChange = stripLineBreaks(onValueChange),
         modifier = modifier,
@@ -68,7 +66,7 @@ fun TextFieldOutlinedPassword(
     isRequired: Boolean = false,
     hasError: Boolean = false,
 ) {
-    MaterialOutlinedTextField(
+    Material3OutlinedTextField(
         value = value,
         onValueChange = stripLineBreaks(onValueChange),
         modifier = modifier,
@@ -110,8 +108,8 @@ private fun selectTrailingIcon(
                 stringResource(id = R.string.designsystem_atom_password_textfield_show_password)
             }
 
-            IconButton(onClick = onClick) {
-                Icon(imageVector = image, contentDescription = description)
+            Material3IconButton(onClick = onClick) {
+                Material3Icon(imageVector = image, contentDescription = description)
             }
         }
     } else {
@@ -131,50 +129,3 @@ private fun selectVisualTransformation(
 }
 
 private fun isShowPasswordAllowed(isEnabled: Boolean, isPasswordVisible: Boolean) = isEnabled && isPasswordVisible
-
-@Preview(showBackground = true)
-@Composable
-internal fun PasswordTextFieldOutlinedPreview() {
-    PreviewWithThemes {
-        TextFieldOutlinedPassword(
-            value = "Input text",
-            onValueChange = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun TextFieldOutlinedPasswordWithLabelPreview() {
-    PreviewWithThemes {
-        TextFieldOutlinedPassword(
-            value = "Input text",
-            label = "Label",
-            onValueChange = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun TextFieldOutlinedPasswordDisabledPreview() {
-    PreviewWithThemes {
-        TextFieldOutlinedPassword(
-            value = "Input text",
-            onValueChange = {},
-            isEnabled = false,
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun TextFieldOutlinedPasswordErrorPreview() {
-    PreviewWithThemes {
-        TextFieldOutlinedPassword(
-            value = "Input text",
-            onValueChange = {},
-            hasError = true,
-        )
-    }
-}
