+++ core/ui/compose/designsystem/src/main/kotlin/app/k9mail/core/ui/compose/designsystem/atom/textfield/TextFieldOutlinedSelect.kt
@@ -15,12 +15,9 @@ import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.withStyle
-import androidx.compose.ui.tooling.preview.Preview
 import app.k9mail.core.ui.compose.designsystem.atom.icon.Icon
 import app.k9mail.core.ui.compose.designsystem.atom.icon.Icons
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
 import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.persistentListOf
 import androidx.compose.material.OutlinedTextField as MaterialOutlinedTextField
 
 @Suppress("LongParameterList")
@@ -147,30 +144,5 @@ private fun <T> DropDownMenu(
                 }
             }
         }
-    }
-}
-
-@Preview
-@Composable
-internal fun TextFieldOutlinedSelectPreview() {
-    PreviewWithThemes {
-        TextFieldOutlinedSelect(
-            options = persistentListOf("Option 1", "Option 2", "Option 3"),
-            selectedOption = "Option 1",
-            onValueChange = {},
-        )
-    }
-}
-
-@Preview
-@Composable
-internal fun TextFieldOutlinedSelectPreviewWithLabel() {
-    PreviewWithThemes {
-        TextFieldOutlinedSelect(
-            options = persistentListOf("Option 1", "Option 2", "Option 3"),
-            selectedOption = "Option 1",
-            onValueChange = {},
-            label = "Label",
-        )
     }
 }
